requirements.txt
feedparser
pandas
streamlit
google-genai
beautifulsoup4
requests